+++ sdks/outpost-python/RELEASES.md
@@ -99,3 +99,13 @@ Based on:
 - [python v0.9.0] sdks/outpost-python
 ### Releases
 - [PyPI v0.9.0] https://pypi.org/project/outpost_sdk/0.9.0 - sdks/outpost-python
+
+## 2026-04-21 19:54:42
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.9.1] sdks/outpost-python
+### Releases
+- [PyPI v0.9.1] https://pypi.org/project/outpost_sdk/0.9.1 - sdks/outpost-python